Proyecto final
Actualizado a la sesiòn del dìa 09/12

+++ proyecto_escuela/src/main/java/org/bedu/proyecto_escuela/service/AlumnoService.java
@@ -2,6 +2,8 @@ package org.bedu.proyecto_escuela.service;
 
 import org.bedu.proyecto_escuela.dto.AlumnoDTO;
 import org.bedu.proyecto_escuela.dto.CreateAlumnoDTO;
+import org.bedu.proyecto_escuela.dto.UpdateAlumnoDTO;
+import org.bedu.proyecto_escuela.exception.AlumnoNotFoundException;
 import org.bedu.proyecto_escuela.mapper.AlumnoMapper;
 import org.bedu.proyecto_escuela.model.Alumno;
 import org.bedu.proyecto_escuela.repository.AlumnoRepository;
@@ -33,23 +35,33 @@ public class AlumnoService {
         return mapper.toDTO(entity);
     }
 
-    public AlumnoDTO update(Long id, CreateAlumnoDTO data) {
+    public void update(Long id, UpdateAlumnoDTO data) throws AlumnoNotFoundException{
         Optional<Alumno> alumnoExistente = repository.findById(id);
         if (alumnoExistente.isPresent()) {
             Alumno alumnoActual = alumnoExistente.get();
+            /*
             alumnoActual.setMatricula(data.getMatricula());
             alumnoActual.setNombre_alumno(data.getNombre_alumno());
             alumnoActual.setSexo(data.getSexo());
             alumnoActual.setTelefono(data.getTelefono());
             alumnoActual.setDireccion(data.getDireccion());
             alumnoActual.setEmail(data.getEmail());
-            return mapper.toDTO(repository.save(alumnoActual));
+            */
+            mapper.update(alumnoActual, data);
+            repository.save(alumnoActual);
         }
-
-        return null;
+        else {
+            throw new AlumnoNotFoundException(id);
+        }
     }
 
-    public void delete(Long id) {
-        repository.deleteById(id);
+    public void delete(Long id) throws AlumnoNotFoundException{
+        Optional<Alumno> alumnoExistente = repository.findById(id);
+        if (alumnoExistente.isPresent()) {
+            repository.deleteById(id);
+        }
+        else {
+            throw new AlumnoNotFoundException(id);
+        }
     }
 }
